Turn off pwm-led lightbar when device sleeps
Leds that support breathe will continue to breathe when asleep, ones
that do not support state or mode will simply turn off.

To support this, the current brightness of the power led is saved,
so it can be restored when the device is woke.

Since sleep/wake calls into the into the backlight path, support
for backlight must be declared even if it isn't present so the
power led state can be set.

This also removes the power led brightness change in the buttons
path. It was interfering with the power led control in the brightness
path.

Change-Id: I1a95185295f4ed56e16a5e4630fafafced9e8583